logged in user can add,edit,delete; Countries has foreignIDs in airport & airline db; airport & airline country input made into dropdown select; authextends main;

+++ app/Http/Controllers/AirlineController.php
@@ -5,10 +5,14 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreAirlineRequest;
 use App\Http\Requests\UpdateAirlineRequest;
 use App\Models\Airline;
+use App\Models\Country;
 use Illuminate\Support\Facades\Gate;
 
 class AirlineController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -26,7 +30,8 @@ class AirlineController extends Controller
      */
     public function create()
     {
-        return view('pages.airline.add_airline');
+        $country = Country::All();
+        return view('pages.airline.add_airline' , compact('country'));
     }
 
     /**
@@ -37,17 +42,22 @@ class AirlineController extends Controller
      */
     public function store(StoreAirlineRequest $request)
     {
+
         $validate = $request->validate([
             'airline_name' => 'required',
             'country_name' => 'required',
             'country_ISO' => 'required',
+            'country_id' => 'required',
         ]);
 
         Airline::create([
             'airline_name' =>request('airline_name'),
             'country_name' =>request('country_name'),
             'country_ISO' =>request('country_ISO'),
+            'country_id' =>request('country_id'),
         ]);
+
+
         return redirect('/show_airline');
     }
 
@@ -72,10 +82,11 @@ class AirlineController extends Controller
      */
     public function edit(Airline $airline)
     {
-        // if(Gate::denies('edit_airline', $airline)){
-        //     return view('pages.denies');
-        // }
-        return view('pages.airline.edit_airline', compact('airline'));
+        if(Gate::denies('edit_airline', $airline)){
+            return view('pages.denies');
+        }
+        $country = Country::All();
+        return view('pages.airline.edit_airline', compact('airline', 'country'));
     }
 
     /**
@@ -99,9 +110,9 @@ class AirlineController extends Controller
      */
     public function destroy(Airline $airline)
     {
-        // if(Gate::denies('delete_airline', $airline)){
-        //     return view('pages.denies');
-        // }
+        if(Gate::denies('delete_airline', $airline)){
+            return view('pages.denies');
+        }
         $airline->delete();
 
         return redirect('/show_airline');
